feat(#80): 채팅내역 API 위치 변경
