+++ src/apis/chatApi/chat.ts
@@ -20,7 +20,7 @@ const CHAT = {
   /** 채팅룸 전체조회 api */
   async getChatRoomList(): Promise<any> {
     const result: AxiosResponse<any> = await instance.get(`${CHAT.path}`);
-    console.log(result.data);
+    // console.log(result.data);
     return result.data.contents;
   },
