+++ src/recoil/atoms/SelectedRoomIdAtom.ts
@@ -1,0 +1,6 @@
+import { atom } from 'recoil';
+
+export const SelectedRoomIdAtom = atom<string>({
+  key: 'selectedRoom',
+  default: '',
+});
